go schemabuilder: Move function parsing to it's own file
